requirements.txt
Django>=5.2,<6.0
Pillow>=10.3
djangorestframework>=3.15.2,<4
django-filter>=24.3,<25
celery>=5.4,<6
redis>=5.0,<6
python-dotenv>=1.0,<2
whitenoise>=6.7,<7
django-allauth>=0.63
requests>=2.32
requests-oauthlib>=2.0
cryptography>=42,<44
python-slugify>=8.0,<9
mysqlclient>=2.2,<3
gunicorn>=21.2,<22
django-celery-beat>=2.5,<3
channels>=4.0,<5
channels-redis>=4.1,<5
daphne>=4.0,<5
